Change app ids from uk.me.jeremygreen.merging to uk.me.jeremygreen.merging2.
