--- app/src/main/java/uk/me/jeremygreen/merging/model/CoordinateDao.kt
@@ -1,17 +1,0 @@
-package uk.me.jeremygreen.merging.model
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-
-@Dao
-internal interface CoordinateDao {
-
-    @Query("SELECT * from coordinates WHERE faceId = :faceId")
-    suspend fun findById(faceId: Long): List<Coordinate>
-
-    @Insert
-    suspend fun addAll(coordinates: List<Coordinate>)
-
-}
-
