--- app/src/main/java/uk/me/jeremygreen/merging/model/ProcessingStage.kt
@@ -1,6 +1,0 @@
-package uk.me.jeremygreen.merging.model
-
-object ProcessingStage {
-    const val unprocessed = 1000
-    const val facesDetected = 2000
-}
