+++ app/src/main/java/uk/me/jeremygreen/merging2/main/ScreenFragmentFactory.kt
@@ -1,0 +1,17 @@
+package uk.me.jeremygreen.merging2.main
+
+internal interface ScreenFragmentFactory<T: ScreenFragment> {
+
+    /**
+     * The ViewPager2 (RecyclerView) id. Images use non-negative Ints, and ViewPager2 uses -1.
+     */
+    val id: Long
+
+    fun createInstance(): T
+
+    /**
+     * Firebase Analytics "screen name".
+     */
+    fun screenName(): String
+
+}
