+++ app/src/main/java/uk/me/jeremygreen/merging2/main/screen/AddImageFragment.kt
@@ -1,0 +1,28 @@
+package uk.me.jeremygreen.merging2.main.screen
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import uk.me.jeremygreen.merging2.R
+import uk.me.jeremygreen.merging2.main.ScreenFragment
+import uk.me.jeremygreen.merging2.main.ScreenFragmentFactory
+
+internal class AddImageFragment : ScreenFragment() {
+
+    companion object: ScreenFragmentFactory<AddImageFragment> {
+        override val id: Long = -2
+        override fun createInstance(): AddImageFragment {
+            return AddImageFragment()
+        }
+        override fun screenName(): String = "AddImage"
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.add_image_screen, container, false)
+    }
+
+}
