+++ app/src/main/java/uk/me/jeremygreen/merging2/model/Onboarding.kt
@@ -1,0 +1,15 @@
+package uk.me.jeremygreen.merging2.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+/**
+ * This table stores a set of accepted onboarding-text versions.
+ */
+@Entity(tableName = "onboarding")
+internal data class Onboarding(
+
+    @PrimaryKey
+    val version: Long
+
+)
