+++ app/src/main/java/uk/me/jeremygreen/merging2/model/OnboardingDao.kt
@@ -1,0 +1,18 @@
+package uk.me.jeremygreen.merging2.model
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+
+@Dao
+internal interface OnboardingDao {
+
+    @Query("SELECT * from onboarding`` WHERE version = :version")
+    suspend fun findById(version: Long): Onboarding?
+
+    @Insert(onConflict = OnConflictStrategy.IGNORE)
+    suspend fun add(version: Onboarding)
+
+}
+
