+++ app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="uk.me.jeremygreen.merging.main.MainActivity">
+      tools:context="uk.me.jeremygreen.merging2.main.MainActivity">
     <item android:id="@+id/actionLicences"
           android:title="@string/actionLicences"
           android:orderInCategory="100"
