+++ app/src/test/java/uk/me/jeremygreen/merging2/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package uk.me.jeremygreen.merging2
+
+import org.junit.Test
+
+import org.junit.Assert.assertEquals
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+internal class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
